textfuly/requirements.txt
amqp==5.1.1
asgiref==3.5.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
blis==0.7.8
bs4==0.0.1
catalogue==2.0.8
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
comtypes==1.1.14
confection==0.0.3
cryptography==38.0.1
cymem==2.0.6
defusedxml==0.7.1
Django==4.0.8
django-celery-beat==2.3.0
django-crispy-forms==1.14.0
django-markdown2==0.3.1
django-timezone-field==5.0
docx2txt==0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
gTTS==2.2.4
idna==3.4
Jinja2==3.1.2
kombu==5.2.4
langcodes==3.3.0
markdown2==2.4.5
MarkupSafe==2.1.1
murmurhash==1.0.8
numpy==1.23.3
oauthlib==3.2.1
packaging==21.3
pathy==0.6.2
pdfminer==20191125
pdfminer.six==20220524
Pillow==9.2.0
preshed==3.0.7
prompt-toolkit==3.0.31
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.9.2
PyJWT==2.5.0
pyparsing==3.0.9
pypiwin32==223
pytesseract==0.3.10
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pyttsx3==2.90
pytz==2022.4
pywin32==304
razorpay==1.3.0
reportlab==3.6.11
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
smart-open==5.2.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
sqlparse==0.4.3
srsly==2.4.4
thinc==8.1.3
tqdm==4.64.1
typer==0.4.2
typing_extensions==4.4.0
tzdata==2022.4
urllib3==1.26.12
vine==5.0.0
wasabi==0.10.1
wcwidth==0.2.5
youtube-transcript-api==0.4.4
